requirements.txt
streamlit
numpy
scikit-learn
joblib
nltk
googlesearch-python
openai
